os/bluestore:  use non-buffered IO for WAL files

+++ src/os/bluestore/BlueFS.h
@@ -695,7 +695,7 @@ private:
   /* signal replay log to include h->file in nearest log flush */
   int _signal_dirty_to_log_D(FileWriter *h);
   int _flush_range_F(FileWriter *h, uint64_t end); // flush up to offset 'end'
-  int _flush_data(FileWriter *h, uint64_t end, bool buffered);
+  uint64_t _flush_data(FileWriter *h, uint64_t end, bool buffered);
   int _flush_F(FileWriter *h, bool force, bool *flushed = nullptr);
   int _flush_envelope_F(FileWriter *h);
   int _fsync(FileWriter *h, bool force_dirty);
